requirements.txt
anyio==4.12.1
python-jose[cryptography]
passlib[bcrypt]
charset-normalizer==3.4.7
fastapi==0.128.8
faster-whisper==1.2.1
httpx==0.28.1
pydantic==2.12.5
pydantic_core==2.41.5
python-multipart==0.0.20
requests==2.32.5
starlette==0.49.3
urllib3==2.6.3
uvicorn==0.39.0
vaderSentiment==3.3.2
Pillow>=10.0.0
